requirements.txt
print-color
vosk
sounddevice
pvporcupine
pyaudio
events
pygame
pyalsaaudio
pyusb
python-vlc